Move LinkedIn and MicrosoftTranslator to Microsoft.Toolkit.Services project
