+++ .gitignore
@@ -223,3 +223,4 @@ project.lock.json
 msbuild.binlog
 *.project.lock.json
 /build/tools
+/Microsoft.Toolkit.Services/Services/MicrosoftTranslator
