+++ Microsoft.Toolkit.Services/Services/LinkedIn/LinkedInDataConfig.cs
@@ -2,16 +2,16 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-namespace Microsoft.Toolkit.Uwp.Services.LinkedIn
+namespace Microsoft.Toolkit.Services.LinkedIn
 {
     /// <summary>
-    /// Strongly typed SiteStandardProfileRequest class.
+    /// Configuration object for specifying richer query information.
     /// </summary>
-    public class LinkedInProfileRequest
+    public class LinkedInDataConfig
     {
         /// <summary>
-        /// Gets or sets url property.
+        /// Gets or sets the query string for filtering service results.
         /// </summary>
-        public string Url { get; set; }
+        public string Query { get; set; }
     }
 }
