+++ Microsoft.Toolkit.Services/Services/LinkedIn/LinkedInPermissions.cs
@@ -4,7 +4,7 @@
 
 using System;
 
-namespace Microsoft.Toolkit.Uwp.Services.LinkedIn
+namespace Microsoft.Toolkit.Services.LinkedIn
 {
     /// <summary>
     /// List of user related data permissions
