+++ Microsoft.Toolkit.Services/Services/LinkedIn/LinkedInProfileRequest.cs
@@ -1,0 +1,17 @@
+// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+namespace Microsoft.Toolkit.Services.LinkedIn
+{
+    /// <summary>
+    /// Strongly typed SiteStandardProfileRequest class.
+    /// </summary>
+    public class LinkedInProfileRequest
+    {
+        /// <summary>
+        /// Gets or sets url property.
+        /// </summary>
+        public string Url { get; set; }
+    }
+}
