+++ Microsoft.Toolkit.Uwp.SampleApp/SamplePages/Microsoft Translator Service/MicrosoftTranslatorPage.xaml.cs
@@ -6,7 +6,7 @@ using System;
 using System.Collections.Generic;
 using System.Globalization;
 using System.Linq;
-using Microsoft.Toolkit.Uwp.Services.MicrosoftTranslator;
+using Microsoft.Toolkit.Services.MicrosoftTranslator;
 using Windows.System;
 using Windows.UI.Xaml;
